backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg>=0.30.0
aiosqlite==0.20.0
anthropic>=0.42.0
claude-agent-sdk>=0.1.51
modal>=1.2.0
python-multipart==0.0.12
python-dotenv==1.0.1
sse-starlette==2.1.0
pydantic>=2.9.0
pydantic-settings>=2.0.0
pyyaml>=6.0
boto3>=1.35.0
duckdb>=1.0.0
pyarrow>=14.0.0
pandas>=2.0.0
pytest>=9.0.0
pytest-asyncio>=1.3.0
httpx>=0.28.0
